Responsive sidebars and new responsive sidebar template
