+++ tailwind.config.ts
@@ -1,5 +1,6 @@
 import { ThemeColors, nextui } from "@nextui-org/react";
 import type { Config } from "tailwindcss";
+import plugin from "tailwindcss/plugin";
 
 const config: Config = {
   content: [
@@ -13,55 +14,65 @@ const config: Config = {
     extend: {},
   },
   darkMode: "class",
-  plugins: [nextui({
-    themes: {
-      dark: {
-        colors: {
-          primary: {
-            50: "#e6f3ff",
-            100: "#c0d7f1",
-            200: "#98bce5",
-            300: "#71a1db",
-            400: "#4b86d1",
-            500: "#336db7",
-            600: "#27558f",
-            700: "#1b3c67",
-            800: "#0e243f",
-            900: "#020c19",
-            DEFAULT: "#112D4E",
-            foreground: "#F8F8F8",
-          },
-          secondary: {
-            DEFAULT: "#1b3c67",
-            foreground: "#F8F8F8",
-          },
-          accent: "#F7418F"
-        } as Partial<ThemeColors & { accent: string }>,
-      },
-      light: {
-        colors: {
-          primary: {
-            50: "#e6f3ff",
-            100: "#c0d7f1",
-            200: "#98bce5",
-            300: "#71a1db",
-            400: "#4b86d1",
-            500: "#336db7",
-            600: "#27558f",
-            700: "#1b3c67",
-            800: "#0e243f",
-            900: "#020c19",
-            DEFAULT: "#112D4E",
-            foreground: "#F8F8F8",
-          },
-          secondary: {
-            DEFAULT: "#1b3c67",
-            foreground: "#F8F8F8",
-          },
-          accent: "#F7418F"
-        } as Partial<ThemeColors & { accent: string }>,
+  plugins: [
+    plugin(({ addUtilities }) => {
+      addUtilities({
+        '.hide-children > *': {
+          display: 'none',
+        },
+      })
+    }),
+
+    nextui({
+      themes: {
+        dark: {
+          colors: {
+            primary: {
+              50: "#e6f3ff",
+              100: "#c0d7f1",
+              200: "#98bce5",
+              300: "#71a1db",
+              400: "#4b86d1",
+              500: "#336db7",
+              600: "#27558f",
+              700: "#1b3c67",
+              800: "#0e243f",
+              900: "#020c19",
+              DEFAULT: "#112D4E",
+              foreground: "#F8F8F8",
+            },
+            secondary: {
+              DEFAULT: "#1b3c67",
+              foreground: "#F8F8F8",
+            },
+            accent: "#F7418F"
+          } as Partial<ThemeColors & { accent: string }>,
+        },
+        light: {
+          colors: {
+            primary: {
+              50: "#e6f3ff",
+              100: "#c0d7f1",
+              200: "#98bce5",
+              300: "#71a1db",
+              400: "#4b86d1",
+              500: "#336db7",
+              600: "#27558f",
+              700: "#1b3c67",
+              800: "#0e243f",
+              900: "#020c19",
+              DEFAULT: "#112D4E",
+              foreground: "#F8F8F8",
+            },
+            secondary: {
+              DEFAULT: "#1b3c67",
+              foreground: "#F8F8F8",
+            },
+            accent: "#F7418F"
+          } as Partial<ThemeColors & { accent: string }>,
+        }
       }
-    }
-  })],
+    })
+  ],
 };
 export default config;
